requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
newsapi-python==0.2.7
python-dotenv==1.0.0
gunicorn==21.2.0
